Move contained classes/struct into own files
