+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/SelectionUIService.HitTestInfo.cs
@@ -1,0 +1,72 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms.Design
+{
+    internal sealed partial class SelectionUIService
+    {
+        private struct HitTestInfo : IEquatable<HitTestInfo>
+        {
+            public readonly int hitTest;
+            public readonly SelectionUIItem selectionUIHit;
+            public readonly bool containerSelector;
+
+            public HitTestInfo(int hitTest, SelectionUIItem selectionUIHit)
+            {
+                this.hitTest = hitTest;
+                this.selectionUIHit = selectionUIHit;
+                containerSelector = false;
+            }
+
+            public HitTestInfo(int hitTest, SelectionUIItem selectionUIHit, bool containerSelector)
+            {
+                this.hitTest = hitTest;
+                this.selectionUIHit = selectionUIHit;
+                this.containerSelector = containerSelector;
+            }
+
+            // Standard 'catch all - rethrow critical' exception pattern
+            public override bool Equals(object obj)
+            {
+                try
+                {
+                    HitTestInfo hi = (HitTestInfo)obj;
+                    return Equals(hi);
+                }
+                catch (Exception ex)
+                {
+                    if (ClientUtils.IsCriticalException(ex))
+                    {
+                        throw;
+                    }
+                }
+
+                return false;
+            }
+
+            public bool Equals(HitTestInfo other)
+                => hitTest == other.hitTest
+                    && selectionUIHit == other.selectionUIHit
+                    && containerSelector == other.containerSelector;
+
+            public static bool operator ==(HitTestInfo left, HitTestInfo right)
+            {
+                return left.Equals(right);
+            }
+
+            public static bool operator !=(HitTestInfo left, HitTestInfo right) => !left.Equals(right);
+
+            public override int GetHashCode()
+            {
+                int hash = hitTest | selectionUIHit.GetHashCode();
+                if (containerSelector)
+                {
+                    hash |= 0x10000;
+                }
+
+                return hash;
+            }
+        }
+    }
+}
